Auto focus and eslint.
