+++ src/components/Shared/DataGrid/components/numberfield/index.js
@@ -1,26 +1,6 @@
-import { useGridApiContext } from '@mui/x-data-grid'
-import CustomTextField from 'src/components/Inputs/CustomTextField'
+import edit from './edit'
 
 export default {
   view: ({ value }) => value,
-  edit({ column: { props }, id, field, value }) {
-    const api = useGridApiContext()
-
-    return (
-      <CustomTextField
-        value={value}
-        label={field}
-        type='number'
-        readOnly={props?.readOnly}
-        autoFocus
-        onChange={e => {
-          api.current.setEditCellValue({
-            id,
-            field,
-            value: parseFloat(e.target.value)
-          })
-        }}
-      />
-    )
-  }
+  edit
 }
